Refactor entire camera stack:
Changes include:
- Changing directory organization to split files
- Standardize options system for both stellarHD and exploreHD
- Improve device registry to use O(1) lookup
- Finalize refactor of schema location to models/
- Rewrite asic controls to use RLock
