+++ backend_py/src/routes/preferences.py
@@ -5,9 +5,8 @@ API endpoints for server perferences
 Handles getting and setting preferences
 """
 
+from backend_py.src.models import SavedPreferencesModel
 from fastapi import APIRouter, Request
-
-from src.models import SavedPreferencesModel
 
 from ..schemas import SimpleRequestStatusModel
 from ..services.preferences import PreferencesManager
