+++ backend_py/src/services/cameras/__init__.py
@@ -1,4 +1,3 @@
-from ..preferences.settings_manager import SettingsManager
 from .device_manager import DeviceManager
 from .device_utils import find_device_with_bus_info, list_diff
 from .pwm import SerialPWMController
@@ -7,6 +6,5 @@ __all__ = [
     "DeviceManager",
     "find_device_with_bus_info",
     "list_diff",
-    "SettingsManager",
     "SerialPWMController",
 ]
