+++ backend_py/src/services/cameras/device_utils.py
@@ -5,7 +5,7 @@ Utility functions for device_manager.py, specifically for finding added devices 
  removed devices
 """
 
-from .device import Device
+from .drivers.device import Device
 
 
 def find_device_with_bus_info(devices: list[Device], bus_info: str) -> Device | None:
