+++ backend_py/src/services/cameras/drivers/ehd/__init__.py
@@ -1,0 +1,3 @@
+from .ehd import EHDDevice
+
+__all__ = ["EHDDevice"]
