+++ backend_py/src/services/cameras/drivers/options/__init__.py
@@ -1,0 +1,4 @@
+from .base_option import BaseOption
+from .xu_option import OptionDescriptor, XUOption
+
+__all__ = ["BaseOption", "XUOption", "OptionDescriptor"]
