+++ backend_py/src/services/cameras/drivers/shd/__init__.py
@@ -1,0 +1,3 @@
+from .shd import SHDDevice
+
+__all__ = ["SHDDevice"]
