+++ backend_py/src/services/cameras/drivers/xu.py
@@ -11,11 +11,21 @@ DWE_DEVICE_TAG = 0x9A
 
 
 class Unit(Enum):
-    SYS_ID = 0x03  # In past was 0x02, but was unused. Was a mistake
-    USR_ID = 0x04
+    """
+    SYS_ID: used for camera controls that are asic level
+
+    USR_ID: used for camera controls that are more user facing
+    """
+
+    SYS_ID = 0x03  # Internal controls for the camera's ASIC
+    USR_ID = 0x04  # User controls for the camera
 
 
 class Selector(Enum):
+    """
+    Selectors for the camera extension unit controls
+    """
+
     SYS_ASIC_RW = 0x01
     SYS_FLASH_CTRL = 0x03
     SYS_FRAME_INFO = 0x06
@@ -36,12 +46,21 @@ class Selector(Enum):
 
 
 class Command(Enum):
+    """
+    Commands for the exploreHD extension unit controls
+    """
+
     H264_BITRATE_CTRL = 0x02
     GOP_CTRL = 0x03
     H264_MODE_CTRL = 0x06
 
 
-class StellarRegisterMap(Enum):
+class StellarRegisterMap:
+    """
+    Map of stellar registers for the ASIC,
+    which can be accessed through the SYS_ASIC_RW selector
+    """
+
     REG_AE = 0x1673
     REG_ADDR_H = 0x1674
     REG_ADDR_L = 0x1675
@@ -55,6 +74,10 @@ class StellarRegisterMap(Enum):
 
 
 class StellarSensorMap:
+    """
+    Map of sensor registers for stellar devices, which can be accessed through the ASIC
+    """
+
     SHUTTER_HIGH = 0x3501
     SHUTTER_LOW = 0x3502
     ISO_HIGH = 0x3508
